+++ intervalsets/src/factory.rs
@@ -1,6 +1,6 @@
 pub use intervalsets_core::factory::{
-    traits, Converter, ConvertingFactory, EIFactory, EmptyFactory, FiniteFactory,
-    HalfBoundedFactory, Identity, TryFiniteFactory, TryHalfBoundedFactory, UnboundedFactory,
+    traits, ConvertingFactory, EmptyFactory, FiniteFactory, HalfBoundedFactory, TryFiniteFactory,
+    TryHalfBoundedFactory, UnboundedFactory,
 };
 use intervalsets_core::sets::{EnumInterval, FiniteInterval, HalfInterval};
 
@@ -8,18 +8,18 @@ use crate::bound::FiniteBound;
 use crate::numeric::{Element, Zero};
 use crate::{Interval, IntervalSet, Side};
 
-impl<T: Element> ConvertingFactory<T, Identity> for Interval<T> {
+impl<T: Element> ConvertingFactory<T> for Interval<T> {
     type Output = Self;
     type Error = crate::error::Error;
 }
 
-impl<T: Element> EmptyFactory<T, Identity> for Interval<T> {
+impl<T: Element> EmptyFactory<T> for Interval<T> {
     fn empty() -> Self::Output {
         FiniteInterval::empty().into()
     }
 }
 
-impl<T: Element> TryFiniteFactory<T, Identity> for Interval<T> {
+impl<T: Element> TryFiniteFactory<T> for Interval<T> {
     fn try_finite(lhs: FiniteBound<T>, rhs: FiniteBound<T>) -> Result<Self::Output, Self::Error> {
         FiniteInterval::try_new_or_empty(lhs, rhs)
             .map_err(Into::into)
@@ -27,7 +27,7 @@ impl<T: Element> TryFiniteFactory<T, Identity> for Interval<T> {
     }
 }
 
-impl<T: Element + Zero> TryHalfBoundedFactory<T, Identity> for Interval<T> {
+impl<T: Element + Zero> TryHalfBoundedFactory<T> for Interval<T> {
     fn try_half_bounded(side: Side, bound: FiniteBound<T>) -> Result<Self::Output, Self::Error> {
         HalfInterval::try_new(side, bound)
             .map_err(Into::into)
@@ -35,24 +35,24 @@ impl<T: Element + Zero> TryHalfBoundedFactory<T, Identity> for Interval<T> {
     }
 }
 
-impl<T: Element> UnboundedFactory<T, Identity> for Interval<T> {
+impl<T: Element> UnboundedFactory<T> for Interval<T> {
     fn unbounded() -> Self::Output {
         EnumInterval::Unbounded.into()
     }
 }
 
-impl<T: Element> ConvertingFactory<T, Identity> for IntervalSet<T> {
+impl<T: Element> ConvertingFactory<T> for IntervalSet<T> {
     type Output = Self;
     type Error = crate::error::Error;
 }
 
-impl<T: Element> EmptyFactory<T, Identity> for IntervalSet<T> {
+impl<T: Element> EmptyFactory<T> for IntervalSet<T> {
     fn empty() -> Self::Output {
         IntervalSet::empty()
     }
 }
 
-impl<T: Element> TryFiniteFactory<T, Identity> for IntervalSet<T> {
+impl<T: Element> TryFiniteFactory<T> for IntervalSet<T> {
     fn try_finite(lhs: FiniteBound<T>, rhs: FiniteBound<T>) -> Result<Self::Output, Self::Error> {
         FiniteInterval::try_new_or_empty(lhs, rhs)
             .map_err(Into::into)
@@ -60,7 +60,7 @@ impl<T: Element> TryFiniteFactory<T, Identity> for IntervalSet<T> {
     }
 }
 
-impl<T: Element + Zero> TryHalfBoundedFactory<T, Identity> for IntervalSet<T> {
+impl<T: Element + Zero> TryHalfBoundedFactory<T> for IntervalSet<T> {
     fn try_half_bounded(side: Side, bound: FiniteBound<T>) -> Result<Self::Output, Self::Error> {
         HalfInterval::try_new(side, bound)
             .map_err(Into::into)
@@ -68,127 +68,9 @@ impl<T: Element + Zero> TryHalfBoundedFactory<T, Identity> for IntervalSet<T> {
     }
 }
 
-impl<T: Element> UnboundedFactory<T, Identity> for IntervalSet<T> {
+impl<T: Element> UnboundedFactory<T> for IntervalSet<T> {
     fn unbounded() -> Self::Output {
         EnumInterval::Unbounded.into()
-    }
-}
-
-pub struct IFactory<T, C>(std::marker::PhantomData<(T, C)>);
-
-impl<T, C> ConvertingFactory<T, C> for IFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    type Output = Interval<C::To>;
-    type Error = crate::error::Error;
-}
-
-impl<T, C> EmptyFactory<T, C> for IFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    fn empty() -> Self::Output {
-        Interval::empty()
-    }
-}
-
-impl<T, C> TryFiniteFactory<T, C> for IFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    fn try_finite(
-        lhs: FiniteBound<C::To>,
-        rhs: FiniteBound<C::To>,
-    ) -> Result<Self::Output, Self::Error> {
-        EIFactory::<T, C>::try_finite(lhs, rhs)
-            .map_err(Into::into)
-            .map(Interval::from)
-    }
-}
-
-impl<T, C> TryHalfBoundedFactory<T, C> for IFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element + Zero,
-{
-    fn try_half_bounded(
-        side: Side,
-        bound: FiniteBound<C::To>,
-    ) -> Result<Self::Output, Self::Error> {
-        EIFactory::<T, C>::try_half_bounded(side, bound)
-            .map_err(Into::into)
-            .map(Interval::from)
-    }
-}
-
-impl<T, C> UnboundedFactory<T, C> for IFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    fn unbounded() -> Self::Output {
-        Interval::unbounded()
-    }
-}
-
-pub struct ISFactory<T, C>(std::marker::PhantomData<(T, C)>);
-
-impl<T, C> ConvertingFactory<T, C> for ISFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    type Output = IntervalSet<C::To>;
-    type Error = crate::error::Error;
-}
-
-impl<T, C> EmptyFactory<T, C> for ISFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    fn empty() -> Self::Output {
-        IntervalSet::empty()
-    }
-}
-
-impl<T, C> TryFiniteFactory<T, C> for ISFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    fn try_finite(
-        lhs: FiniteBound<C::To>,
-        rhs: FiniteBound<C::To>,
-    ) -> Result<Self::Output, Self::Error> {
-        IFactory::<T, C>::try_finite(lhs, rhs).map(IntervalSet::from)
-    }
-}
-
-impl<T, C> TryHalfBoundedFactory<T, C> for ISFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element + Zero,
-{
-    fn try_half_bounded(
-        side: Side,
-        bound: FiniteBound<C::To>,
-    ) -> Result<Self::Output, Self::Error> {
-        IFactory::<T, C>::try_half_bounded(side, bound).map(IntervalSet::from)
-    }
-}
-
-impl<T, C> UnboundedFactory<T, C> for ISFactory<T, C>
-where
-    C: Converter<T>,
-    C::To: Element,
-{
-    fn unbounded() -> Self::Output {
-        IntervalSet::unbounded()
     }
 }
 
@@ -198,7 +80,7 @@ mod tests {
 
     #[test]
     fn test_interval_factory() {
-        let a: Interval<_> = EIFactory::<u32, Identity>::closed(0, 10).into();
+        let a = Interval::<u32>::closed(0, 10);
         let b = Interval::<u32>::closed(0, 10);
         assert_eq!(a, b);
     }